+++ R/data.R
@@ -2,7 +2,7 @@
 
 #' @encoding UTF-8
 #' @title NW_Megafires example dataset
-#' @format A \emph{mts_monitor} object with 1080 rows and 141 columns of data.
+#' @format A \emph{mts_monitor} object with 1080 rows and 143 columns of data.
 #' @description The \code{NW_Megafires} dataset provides a quickly loadable
 #' version of a \emph{mts_monitor} object for practicing and code examples.
 #'
@@ -12,15 +12,16 @@
 #' northern Idaho. The NW_Megafires dataset contains monitoring data for the
 #' Pacific Northwest from July 24 through September 06, 2015.
 #'
-#' This dataset was generated on 2022-10-14 by running:
+#' This dataset was generated on 2022-10-28 by running:
 #'
 #' \preformatted{
 #' library(AirMonitor)
 #'
 #' NW_Megafires <-
-#'   airnow_loadAnnual(2015) \%>\%
-#'   monitor_filterMeta(stateCode \%in\% c("WA", "OR", "ID")) \%>\%
-#'   monitor_filterDate(20150724, 20150907)
+#'   monitor_loadAnnual(2015, epaPreference = "epa_aqs") %>%
+#'   monitor_filterMeta(stateCode %in% c("WA", "OR", "ID")) %>%
+#'   monitor_filterDate(20150724, 20150907, timezone = "America/Los_Angeles") %>%
+#'   monitor_dropEmpty()
 #'
 #' save(NW_Megafires, file = "data/NW_Megafires.rda")
 #' }